+++ src/components/profile/ProfileForm.js
@@ -1,0 +1,66 @@
+import { getUser, updateUser } from "../../store/Auth";
+import SubmitBtn from "../common/SubmitBtn";
+
+const ProfileForm = () => {
+  const user = getUser();
+  return `
+    <form id="profile-form">
+      <div class="mb-4">
+        <label
+          for="username"
+          class="block text-gray-700 text-sm font-bold mb-2"
+          >사용자 이름</label
+        >
+        <input
+          type="text"
+          id="username"
+          name="username"
+          value="${user.username}"
+          class="w-full p-2 border rounded"
+        />
+      </div>
+      <div class="mb-4">
+        <label
+          for="email"
+          class="block text-gray-700 text-sm font-bold mb-2"
+          >이메일</label
+        >
+        <input
+          type="email"
+          id="email"
+          name="email"
+          value="${user.email}"
+          class="w-full p-2 border rounded"
+        />
+      </div>
+      <div class="mb-6">
+        <label
+          for="bio"
+          class="block text-gray-700 text-sm font-bold mb-2"
+          >자기소개</label
+        >
+        <textarea
+          id="bio"
+          name="bio"
+          rows="4"
+          class="w-full p-2 border rounded"
+        >
+          ${user.bio}
+        </textarea>
+      </div>
+      ${SubmitBtn({ text: "프로필 업데이트" })}
+    </form>
+  `;
+};
+
+document.addEventListener("submit", (e) => {
+  if (e.target.id === "profile-form") {
+    e.preventDefault();
+    const username = document.getElementById("username").value;
+    const email = document.getElementById("email").value;
+    const bio = document.getElementById("bio").value;
+    updateUser({ username, email, bio });
+  }
+});
+
+export default ProfileForm;
