+++ src/pages/Profile.js
@@ -1,10 +1,8 @@
 import Header from "../components/common/Header";
 import Footer from "../components/common/Footer";
-import SubmitBtn from "../components/common/SubmitBtn";
-import { getUser } from "../store/Auth";
-const ProfilePage = () => {
-  const user = getUser();
+import ProfileForm from "../components/profile/ProfileForm";
 
+const ProfilePage = () => {
   return `
   <div id="root">
     <div class="bg-gray-100 min-h-screen flex justify-center">
@@ -16,52 +14,7 @@ const ProfilePage = () => {
             <h2 class="text-2xl font-bold text-center text-blue-600 mb-8">
               내 프로필
             </h2>
-            <form>
-              <div class="mb-4">
-                <label
-                  for="username"
-                  class="block text-gray-700 text-sm font-bold mb-2"
-                  >사용자 이름</label
-                >
-                <input
-                  type="text"
-                  id="username"
-                  name="username"
-                  value="${user.username}"
-                  class="w-full p-2 border rounded"
-                />
-              </div>
-              <div class="mb-4">
-                <label
-                  for="email"
-                  class="block text-gray-700 text-sm font-bold mb-2"
-                  >이메일</label
-                >
-                <input
-                  type="email"
-                  id="email"
-                  name="email"
-                  value="${user.email}"
-                  class="w-full p-2 border rounded"
-                />
-              </div>
-              <div class="mb-6">
-                <label
-                  for="bio"
-                  class="block text-gray-700 text-sm font-bold mb-2"
-                  >자기소개</label
-                >
-                <textarea
-                  id="bio"
-                  name="bio"
-                  rows="4"
-                  class="w-full p-2 border rounded"
-                >
-                  ${user.bio}
-                </textarea>
-              </div>
-              ${SubmitBtn({ text: "프로필 업데이트" })}
-            </form>
+            ${ProfileForm()}
           </div>
         </main>
